git-svn-id: svn://fb09-pasig.umwelt.uni-giessen.de/cmf/trunk@61 a735f7ef-1525-764e-8607-acac38503800

+++ cmf/cmf_core_src/math/timeseries.h
@@ -85,7 +85,7 @@ namespace cmf {
 			/// Last date of measurements
 			cmf::math::Time end()	const		{return begin()+step()*int(m_data->values.size());}
 			/// Method for the interpolation (0 - Nearest neighbor, 1- linear, 2 - cubic spline (not implemented yet)
-			double interpolationpower() const {return m_data->interpolationpower;}
+			int interpolationpower() const {return m_data->interpolationpower;}
 			/// Appends a measurement
 			void add(double Value);
 			/// returns true if no values are added to the timeseries
